Time: 3 ms (28.07%), Space: 40.6 MB (56.14%) - LeetHub

+++ 1064-smallest-integer-divisible-by-k/1064-smallest-integer-divisible-by-k.java
@@ -1,6 +1,6 @@
 class Solution {
     public int smallestRepunitDivByK(int k) {
-        long n=0;
+        int n=0;
         if(k%2==0){
             return -1;
         }
